# central-command/server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ogbenjuwa Central Command API",
    "description": "Community safety platform — alert management, patrol tracking, user administration, and real-time incident response for the Idoma Region, Benue State.",
    "version": "1.0.0",
    "contact": { "name": "Ogbenjuwa Team" }
  },
  "servers": [
    { "url": "http://localhost:4001/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "type": "string" },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "lgaId": { "type": "string", "format": "uuid" },
          "location": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "investigating", "resolved", "false_alarm"] },
          "isPublic": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/alerts": {
      "get": {
        "summary": "List alerts (paginated)",
        "tags": ["Alerts"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Paginated alerts list", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Alert" } } } } } } } }
      },
      "post": {
        "summary": "Create a new alert",
        "tags": ["Alerts"],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Alert" } } } },
        "responses": { "201": { "description": "Alert created" } }
      }
    },
    "/alerts/{id}": {
      "get": {
        "summary": "Get alert by ID",
        "tags": ["Alerts"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Alert details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Update alert",
        "tags": ["Alerts"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Alert updated" } }
      }
    },
...
  }
}